detectProject/detect_requirements.txt
Django==4.2.11
django-cors-headers==4.3.1
sqlparse==0.4.4
uwsgi==2.0.25.1
gunicorn==22.0.0
python-dotenv==1.0.1
ultralytics==8.1.45
pillow==10.3.0
